+++ Calculadora/C_functions/math_em_c.h
@@ -1,4 +1,4 @@
-int factorial(int);
+double factorial(double);
 double plus(double, double);
 double minus(double, double);
 double multiply(double, double);
